+++ tests/test_auth.py
@@ -3,6 +3,7 @@ import mock
 import pytest
 
 from httpie.plugins.builtin import HTTPBasicAuth
+from httpie.status import ExitStatus
 from httpie.utils import ExplicitNullAuth
 from utils import http, add_auth, HTTP_OK, MockEnvironment
 import httpie.cli.constants
@@ -14,6 +15,7 @@ def test_basic_auth(httpbin_both):
              'GET', httpbin_both + '/basic-auth/user/password')
     assert HTTP_OK in r
     assert r.json == {'authenticated': True, 'user': 'user'}
+
 
 @pytest.mark.parametrize('argument_name', ['--auth-type', '-A'])
 def test_digest_auth(httpbin_both, argument_name):
@@ -77,6 +79,8 @@ def test_missing_auth(httpbin):
 
 
 def test_netrc(httpbin_both):
+    # This one gets handled by requests (no --auth, --auth-type present),
+    # that’s why we patch inside `requests.sessions`.
     with mock.patch('requests.sessions.get_netrc_auth') as get_netrc_auth:
         get_netrc_auth.return_value = ('httpie', 'password')
         r = http(httpbin_both + '/basic-auth/httpie/password')
@@ -85,19 +89,11 @@ def test_netrc(httpbin_both):
 
 
 def test_ignore_netrc(httpbin_both):
-    with mock.patch('requests.sessions.get_netrc_auth') as get_netrc_auth:
+    with mock.patch('httpie.cli.argparser.get_netrc_auth') as get_netrc_auth:
         get_netrc_auth.return_value = ('httpie', 'password')
         r = http('--ignore-netrc', httpbin_both + '/basic-auth/httpie/password')
         assert get_netrc_auth.call_count == 0
         assert 'HTTP/1.1 401 UNAUTHORIZED' in r
-
-
-def test_ignore_netrc_null_auth():
-    args = httpie.cli.definition.parser.parse_args(
-        args=['--ignore-netrc', 'example.org'],
-        env=MockEnvironment(),
-    )
-    assert isinstance(args.auth, ExplicitNullAuth)
 
 
 def test_ignore_netrc_together_with_auth():
@@ -107,10 +103,40 @@ def test_ignore_netrc_together_with_auth():
     )
     assert isinstance(args.auth, HTTPBasicAuth)
 
-def test_honor_netrc_authtype(httpbin_both):
-    with mock.patch('requests.sessions.get_netrc_auth') as get_netrc_auth:
-        get_netrc_auth.return_value = ('httpie', 'password')
-        r = http('--auth-type=basic','GET', httpbin_both + '/basic-auth/httpie/password')
-        assert get_netrc_auth.call_count == 1
-        assert HTTP_OK in r
 
+def test_ignore_netrc_with_auth_type_resulting_in_missing_auth(httpbin):
+    with mock.patch('httpie.cli.argparser.get_netrc_auth') as get_netrc_auth:
+        get_netrc_auth.return_value = ('httpie', 'password')
+        r = http(
+            '--ignore-netrc',
+            '--auth-type=basic',
+            httpbin + '/basic-auth/httpie/password',
+            tolerate_error_exit_status=True,
+        )
+    assert get_netrc_auth.call_count == 0
+    assert r.exit_status == ExitStatus.ERROR
+    assert '--auth required' in r.stderr
+
+
+@pytest.mark.parametrize(
+    argnames=['auth_type', 'endpoint'],
+    argvalues=[
+        ('basic', '/basic-auth/httpie/password'),
+        ('digest', '/digest-auth/auth/httpie/password'),
+    ],
+)
+def test_auth_plugin_netrc_parse(auth_type, endpoint, httpbin):
+    # Test
+    with mock.patch('httpie.cli.argparser.get_netrc_auth') as get_netrc_auth:
+        get_netrc_auth.return_value = ('httpie', 'password')
+        r = http('--auth-type', auth_type, httpbin + endpoint)
+    assert get_netrc_auth.call_count == 1
+    assert HTTP_OK in r
+
+
+def test_ignore_netrc_null_auth():
+    args = httpie.cli.definition.parser.parse_args(
+        args=['--ignore-netrc', 'example.org'],
+        env=MockEnvironment(),
+    )
+    assert isinstance(args.auth, ExplicitNullAuth)