+++ CHANGELOG.rst
@@ -15,6 +15,8 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 * Added support for ``$XDG_CONFIG_HOME`` (`#920`_).
 * Added support for custom content types for uploaded files (`#668`_).
 * Added support for ``Set-Cookie``-triggered cookie expiration (`#853`_).
+* Added ``netrc`` support for auth plugins.
+  Enabled for ``--auth-type=basic`` and ``digest``, 3rd parties may opt in (`#718`_, `#719`_, `#852`_, `#934`_).
 
 
 `2.1.0`_ (2020-04-18)
@@ -439,9 +441,13 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 .. _#128: https://github.com/jakubroztocil/httpie/issues/128
 .. _#488: https://github.com/jakubroztocil/httpie/issues/488
 .. _#668: https://github.com/jakubroztocil/httpie/issues/668
+.. _#718: https://github.com/jakubroztocil/httpie/issues/718
+.. _#719: https://github.com/jakubroztocil/httpie/issues/719
 .. _#840: https://github.com/jakubroztocil/httpie/issues/840
 .. _#853: https://github.com/jakubroztocil/httpie/issues/853
+.. _#852: https://github.com/jakubroztocil/httpie/issues/852
 .. _#870: https://github.com/jakubroztocil/httpie/issues/870
 .. _#895: https://github.com/jakubroztocil/httpie/issues/895
 .. _#920: https://github.com/jakubroztocil/httpie/issues/920
 .. _#925: https://github.com/jakubroztocil/httpie/issues/925
+.. _#934: https://github.com/jakubroztocil/httpie/issues/934